sonic and shredderdrone reworks
sonic's acceleration gimmick didnt feel that noticeable so i upped it and removed the damage falloff because getting the most out of the bullet's acceleration requires pushing forward during shooting. For pounder-esque tanks, that can easily be really risky to do and i think that risk deserves a bit of reward :P

as for shredderdrone, the extreme bullet health buff and damage nerf is so that it feels more like sawblades. i also shrunk the saw's FOV a lot and made it skynet so that it's heat seeking was more competent and more consistently in-line with what the player is actually aiming at